requirements.txt
joblib==1.4.2
mplfinance==0.12.10b0
numpy==2.0.2
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
scikit-learn==1.6.1
streamlit==1.44.0
ta==0.11.0
gnews==0.4.1
huggingface-hub==0.29.3
joblib==1.4.2
torch==2.6.0
transformers==4.50.2